Fix Android Bug
So some part of Android (Chrome, Gboard, who knows?) changes the "enter" keypress to a "next" keypress if there are more editable input fields lower down on the page. This "next" keypress doesn't seem to be able to be captured so I can't grab it like I'm doing for the "enter" keypress. The only way to fix this that I've found currently that doesn't totally break the page is to make the Pokemon Position input the last editable input on the page and thus get the "enter" keypress when in that input field. Super dumb and makes me change my UI to get around this stupid "feature" that can't be changed.