Hack for "next" button
So as I was ranting about the stupid problem mentioned in the last commit, I had an idea. Since I'm not actually submitting anything together with these "form" elements, they don't actually need to be part of the same form. So would the "next" button go away if there were no more editable inputs within the current "form"? The answer is yes apparently. Seems you get the "enter"/"done"/"go" button whenever you're in the last input field of the particular form you're in. So if you only have one editable field in that particular form, it'll always have the "enter" button.

The UI is once again saved.